rename utils/posts.server.tsx
