+++ app/routes/index.tsx
@@ -1,7 +1,7 @@
 import { Link, LinksFunction, LoaderFunction, useLoaderData } from "remix";
 import indexStylesUrl from "~/styles/index.css";
 import folder from "~/images/home_folder.png";
-import { getPosts } from "~/utils/postsFromDb";
+import { getPosts } from "~/utils/posts.server";
 import { Post } from "~/types/Post";
 
 export const links: LinksFunction = () => {
